Alternative for logging location changes.
I think this won't log as much information as `save("location")` but it has the advantage of being async inside OpenSIPS, so potentially much faster than using the DB layer and going through the Node.js service.
To enable it you must set the `usrloc_db_mode` to 0 and `nofity_register` to a true value.